New User-Comment relation

+++ app/controllers/comments_controller.rb
@@ -1,5 +1,7 @@
 class CommentsController < ApplicationController
   def new
+    session_notice(:danger, 'You must be logged in!', login_path) unless logged_in?
+
     @article = Article.find(params[:article_id])
     @comment = @article.comments.build
   end
@@ -7,6 +9,7 @@ class CommentsController < ApplicationController
   def create
     @article = Article.find(params[:article_id])
     @comment = @article.comments.build(comment_params)
+    @comment.user = current_user
 
     if @comment.save
       redirect_to @article
@@ -16,7 +19,14 @@ class CommentsController < ApplicationController
   end
 
   def edit
+    session_notice(:danger, 'You must be logged in!', login_path) unless logged_in?
+
     @comment = Comment.find(params[:id])
+
+    if logged_in?
+      session_notice(:danger, 'Wrong User') unless equal_with_current_user?(@comment.user)
+    end
+
     @article = @comment.article
   end
 
@@ -32,10 +42,16 @@ class CommentsController < ApplicationController
   end
 
   def destroy
-    comment = Comment.find(params[:id])
-    comment.destroy
+    session_notice(:danger, 'You must be logged in!', login_path) unless logged_in?
 
-    redirect_to comment.article
+    comment = Comment.find(params[:id])
+
+    if equal_with_current_user?(comment.user)
+      comment.destroy
+      redirect_to comment.article
+    else
+      session_notice(:danger, 'Wrong User')
+    end
   end
 
   private
